more refactoring of app.py
