+++ bin/frontend/generate-env-codespace
@@ -6,3 +6,8 @@ template = File.read '/workspaces/aws-bootcamp-cruddur-2023/erb/frontend-react-j
 content = ERB.new(template).result(binding)
 filename = "/workspaces/aws-bootcamp-cruddur-2023/frontend-react-js.env"
 File.write(filename, content)
+
+template = File.read 'erb/sync.env.erb'
+content = ERB.new(template).result(binding)
+filename = "sync.env"
+File.write(filename, content)
