+++ erb/frontend-react-js-codespace.env.erb
@@ -1,5 +1,10 @@
 REACT_APP_BACKEND_URL=https://<%= ENV['CODESPACE_NAME'] %>-4567.<%= ENV['GITHUB_CODESPACES_PORT_FORWARDING_DOMAIN'] %> 
+REACT_APP_FRONTEND_URL=https://<%= ENV['CODESPACE_NAME'] %>-3000.<%= ENV['GITHUB_CODESPACES_PORT_FORWARDING_DOMAIN'] %>
 REACT_APP_AWS_PROJECT_REGION=<%= ENV['AWS_DEFAULT_REGION'] %>
 REACT_APP_AWS_COGNITO_REGION=<%= ENV['AWS_DEFAULT_REGION'] %>
 REACT_APP_AWS_USER_POOL_ID=<%= ENV['AWS_COGNITO_USER_POOL_ID'] %>
 REACT_APP_CLIENT_ID=<%= ENV['AWS_COGNITO_USER_POOL_CLIENT_ID'] %>
+REACT_APP_API_GATEWAY_ENDPOINT_URL=https://pvfc3kkig8.execute-api.us-east-1.amazonaws.com
+
+
+
