+++ erb/sync.env-codespace.erb
@@ -1,0 +1,5 @@
+SYNC_S3_BUCKET=bennieo.me
+SYNC_CLOUDFRONT_DISTRUBTION_ID=E3SZVBPXMZPL1M
+SYNC_BUILD_DIR=<%= ENV['THEIA_WORKSPACE_ROOT'] %>/frontend-react-js/build
+SYNC_OUTPUT_CHANGESET_PATH=<%=  ENV['THEIA_WORKSPACE_ROOT'] %>/tmp/sync-changeset.json
+SYNC_AUTO_APPROVE=false
